requirements.txt
awscli
boto3
botocore
moto
pymysql